Split RemoteServiceActivity into multiple activities
