+++ OpenKeychain/src/main/AndroidManifest.xml
@@ -283,8 +283,7 @@
                 <data android:mimeType="text/pgp" />
 
                 <!--
-                  This links to attached asc files in AOSP mail. It is deactivated because of
-                  https://github.com/open-keychain/open-keychain/issues/290
+                  This links to attached asc files in AOSP Mail
                 -->
                 <data android:mimeType="text/plain" />
             </intent-filter>
@@ -783,7 +782,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.ExitActivity"
+            android:name=".ui.PanicExitActivity"
             android:theme="@android:style/Theme.NoDisplay" />
 
         <!-- Internal services/content providers (not exported) -->
@@ -807,9 +806,21 @@
             android:exported="false"
             android:label="@string/keyserver_sync_settings_title" />
 
-        <!-- Internal classes of the remote APIs (not exported) -->
+        <!-- Internal classes of the remote APIs (not exported!) -->
         <activity
-            android:name=".remote.ui.RemoteServiceActivity"
+            android:name=".remote.ui.RemoteCreateAccountActivity"
+            android:exported="false"
+            android:label="@string/app_name" />
+        <activity
+            android:name=".remote.ui.RemoteErrorActivity"
+            android:exported="false"
+            android:label="@string/app_name" />
+        <activity
+            android:name=".remote.ui.RemoteRegisterActivity"
+            android:exported="false"
+            android:label="@string/app_name" />
+        <activity
+            android:name=".remote.ui.RemoteSelectPubKeyActivity"
             android:exported="false"
             android:label="@string/app_name" />
         <activity
@@ -835,8 +846,7 @@
             android:exported="false" />
 
         <!-- DEPRECATED service,
-           using this service may lead to truncated data being returned to the caller
-         -->
+             using this service may lead to truncated data being returned to the caller -->
         <service
             android:name=".remote.OpenPgpService"
             android:enabled="true"
@@ -864,6 +874,7 @@
         <!-- Contact Sync services -->
         <service
             android:name=".service.DummyAccountService"
+            android:exported="true"
             tools:ignore="ExportedService">
             <intent-filter>
                 <action android:name="android.accounts.AccountAuthenticator" />
@@ -890,6 +901,7 @@
                 android:resource="@xml/sync_adapter_contacts_structure" />
         </service>
 
+        <!-- keyserver sync service -->
         <service
             android:name=".service.KeyserverSyncAdapterService"
             android:exported="true"
@@ -904,7 +916,8 @@
                 android:resource="@xml/sync_adapter_keys" />
         </service>
 
-        <!-- Storage Provider for temporary decrypted files -->
+        <!-- Storage Provider for temporary decrypted files.
+             For security considerations, read class! -->
         <provider
             android:name=".provider.TemporaryFileProvider"
             android:authorities="${applicationId}.tempstorage"
